fix render swarm for local cloud build
